+++ z80.hpp
@@ -105,14 +105,30 @@ class Z80
     // bit table
     const unsigned char bits[8] = {0b00000001, 0b00000010, 0b00000100, 0b00001000, 0b00010000, 0b00100000, 0b01000000, 0b10000000};
     // flag setter
-    inline void setFlagS(bool on) { on ? reg.pair.F |= flagS() : reg.pair.F &= ~flagS(); }
-    inline void setFlagZ(bool on) { on ? reg.pair.F |= flagZ() : reg.pair.F &= ~flagZ(); }
-    inline void setFlagY(bool on) { on ? reg.pair.F |= flagY() : reg.pair.F &= ~flagY(); }
-    inline void setFlagH(bool on) { on ? reg.pair.F |= flagH() : reg.pair.F &= ~flagH(); }
-    inline void setFlagX(bool on) { on ? reg.pair.F |= flagX() : reg.pair.F &= ~flagX(); }
-    inline void setFlagPV(bool on) { on ? reg.pair.F |= flagPV() : reg.pair.F &= ~flagPV(); }
-    inline void setFlagN(bool on) { on ? reg.pair.F |= flagN() : reg.pair.F &= ~flagN(); }
-    inline void setFlagC(bool on) { on ? reg.pair.F |= flagC() : reg.pair.F &= ~flagC(); }
+    inline void setFlagS() { reg.pair.F |= flagS(); }
+    inline void setFlagZ() { reg.pair.F |= flagZ(); }
+    inline void setFlagY() { reg.pair.F |= flagY(); }
+    inline void setFlagH() { reg.pair.F |= flagH(); }
+    inline void setFlagX() { reg.pair.F |= flagX(); }
+    inline void setFlagPV() { reg.pair.F |= flagPV(); }
+    inline void setFlagN() { reg.pair.F |= flagN(); }
+    inline void setFlagC() { reg.pair.F |= flagC(); }
+    inline void resetFlagS() { reg.pair.F &= ~flagS(); }
+    inline void resetFlagZ() { reg.pair.F &= ~flagZ(); }
+    inline void resetFlagY() { reg.pair.F &= ~flagY(); }
+    inline void resetFlagH() { reg.pair.F &= ~flagH(); }
+    inline void resetFlagX() { reg.pair.F &= ~flagX(); }
+    inline void resetFlagPV() { reg.pair.F &= ~flagPV(); }
+    inline void resetFlagN() { reg.pair.F &= ~flagN(); }
+    inline void resetFlagC() { reg.pair.F &= ~flagC(); }
+    inline void setFlagS(bool on) { on ? setFlagS() : resetFlagS(); }
+    inline void setFlagZ(bool on) { on ? setFlagZ() : resetFlagZ(); }
+    inline void setFlagY(bool on) { on ? setFlagY() : resetFlagY(); }
+    inline void setFlagH(bool on) { on ? setFlagH() : resetFlagH(); }
+    inline void setFlagX(bool on) { on ? setFlagX() : resetFlagX(); }
+    inline void setFlagPV(bool on) { on ? setFlagPV() : resetFlagPV(); }
+    inline void setFlagN(bool on) { on ? setFlagN() : resetFlagN(); }
+    inline void setFlagC(bool on) { on ? setFlagC() : resetFlagC(); }
 
     inline void setFlagXY(unsigned char value)
     {
@@ -1815,9 +1831,9 @@ class Z80
         setBC(bc);
         setDE(de);
         setHL(hl);
-        setFlagH(false);
+        resetFlagH();
         setFlagPV(bc != 0);
-        setFlagN(false);
+        resetFlagN();
         unsigned char an = reg.pair.A + n;
         setFlagY(an & 0b00000010);
         setFlagX(an & 0b00001000);
@@ -1995,8 +2011,8 @@ class Z80
     inline void setFlagByRotate(unsigned char n, bool carry, bool isA = false)
     {
         setFlagC(carry);
-        setFlagH(false);
-        setFlagN(false);
+        resetFlagH();
+        resetFlagN();
         setFlagXY(n);
         if (!isA) {
             setFlagS(n & 0x80);
@@ -2983,7 +2999,7 @@ class Z80
     inline void setFlagByIncrement(unsigned char before)
     {
         unsigned char finalResult = before + 1;
-        setFlagN(false);
+        resetFlagN();
         setFlagZ(0 == finalResult);
         setFlagS(0x80 & finalResult);
         setFlagH((finalResult & 0x0F) == 0x00);
@@ -2994,7 +3010,7 @@ class Z80
     inline void setFlagByDecrement(unsigned char before)
     {
         unsigned char finalResult = before - 1;
-        setFlagN(true);
+        setFlagN();
         setFlagZ(0 == finalResult);
         setFlagS(0x80 & finalResult);
         setFlagH((finalResult & 0x0F) == 0x0F);
@@ -3681,7 +3697,7 @@ class Z80
     {
         int result = before + addition;
         int carrybits = before ^ addition ^ result;
-        setFlagN(false);
+        resetFlagN();
         setFlagXY((result & 0xFF00) >> 8);
         setFlagC((carrybits & 0x10000) != 0);
         setFlagH((carrybits & 0x1000) != 0);
@@ -3693,7 +3709,7 @@ class Z80
         int carrybits = before ^ addition ^ result;
         unsigned short finalResult = (unsigned short)(result);
         // same as ADD
-        setFlagN(false);
+        resetFlagN();
         setFlagXY((finalResult & 0xFF00) >> 8);
         setFlagC((carrybits & 0x10000) != 0);
         setFlagH((carrybits & 0x1000) != 0);
@@ -3849,7 +3865,7 @@ class Z80
         int result = before - subtract;
         int carrybits = before ^ subtract ^ result;
         unsigned short finalResult = (unsigned short)result;
-        setFlagN(true);
+        setFlagN();
         setFlagXY((finalResult & 0xFF00) >> 8);
         setFlagC((carrybits & 0x10000) != 0);
         setFlagH((carrybits & 0x1000) != 0);
@@ -3877,33 +3893,35 @@ class Z80
         consumeClock(7);
     }
 
-    inline void setFlagByLogical(bool h)
+    inline void setFlagByLogical()
     {
         setFlagS(reg.pair.A & 0x80);
         setFlagZ(reg.pair.A == 0);
         setFlagXY(reg.pair.A);
-        setFlagH(h);
         setFlagPV(isEvenNumberBits(reg.pair.A));
-        setFlagN(false);
-        setFlagC(false);
+        resetFlagN();
+        resetFlagC();
     }
 
     inline void and8(unsigned char n)
     {
         reg.pair.A &= n;
-        setFlagByLogical(true);
+        setFlagByLogical();
+        setFlagH();
     }
 
     inline void or8(unsigned char n)
     {
         reg.pair.A |= n;
-        setFlagByLogical(false);
+        setFlagByLogical();
+        resetFlagH();
     }
 
     inline void xor8(unsigned char n)
     {
         reg.pair.A ^= n;
-        setFlagByLogical(false);
+        setFlagByLogical();
+        resetFlagH();
     }
 
     // AND Register
@@ -4235,8 +4253,8 @@ class Z80
         if (ctx->isDebug()) ctx->log("[%04X] CPL %s", ctx->reg.PC - 1, ctx->registerDump(0b111));
 #endif
         ctx->reg.pair.A = ~ctx->reg.pair.A;
-        ctx->setFlagH(true);
-        ctx->setFlagN(true);
+        ctx->setFlagH();
+        ctx->setFlagN();
         ctx->setFlagXY(ctx->reg.pair.A);
     }
 
@@ -4259,7 +4277,7 @@ class Z80
         if (ctx->isDebug()) ctx->log("[%04X] CCF <C:%s -> %s>", ctx->reg.PC - 1, ctx->isFlagC() ? "ON" : "OFF", !ctx->isFlagC() ? "ON" : "OFF");
 #endif
         ctx->setFlagH(ctx->isFlagC());
-        ctx->setFlagN(false);
+        ctx->resetFlagN();
         ctx->setFlagC(!ctx->isFlagC());
         ctx->setFlagXY(ctx->reg.pair.A);
     }
@@ -4270,9 +4288,9 @@ class Z80
 #ifndef Z80_DISABLE_DEBUG
         if (ctx->isDebug()) ctx->log("[%04X] SCF <C:%s -> ON>", ctx->reg.PC - 1, ctx->isFlagC() ? "ON" : "OFF");
 #endif
-        ctx->setFlagH(false);
-        ctx->setFlagN(false);
-        ctx->setFlagC(true);
+        ctx->resetFlagH();
+        ctx->resetFlagN();
+        ctx->setFlagC();
         ctx->setFlagXY(ctx->reg.pair.A);
     }
 
@@ -4344,8 +4362,8 @@ class Z80
         setFlagZ(n ? false : true);
         setFlagPV(isFlagZ());
         setFlagS(!isFlagZ() && 7 == bit);
-        setFlagH(true);
-        setFlagN(false);
+        setFlagH();
+        resetFlagN();
         setFlagXY(*rp);
     }
 
@@ -4368,8 +4386,8 @@ class Z80
         setFlagZ(!n);
         setFlagPV(isFlagZ());
         setFlagS(!isFlagZ() && 7 == bit);
-        setFlagH(true);
-        setFlagN(false);
+        setFlagH();
+        resetFlagN();
         setFlagXY((reg.WZ & 0xFF00) >> 8);
     }
 
@@ -4392,8 +4410,8 @@ class Z80
         setFlagZ(!n);
         setFlagPV(isFlagZ());
         setFlagS(!isFlagZ() && 7 == bit);
-        setFlagH(true);
-        setFlagN(false);
+        setFlagH();
+        resetFlagN();
         setFlagXY((reg.WZ & 0xFF00) >> 8);
     }
 
@@ -4416,8 +4434,8 @@ class Z80
         setFlagZ(!n);
         setFlagPV(isFlagZ());
         setFlagS(!isFlagZ() && 7 == bit);
-        setFlagH(true);
-        setFlagN(false);
+        setFlagH();
+        resetFlagN();
         setFlagXY((reg.WZ & 0xFF00) >> 8);
     }
 
@@ -5465,9 +5483,9 @@ class Z80
 #endif
         setFlagS(i & 0x80);
         setFlagZ(i == 0);
-        setFlagH(false);
+        resetFlagH();
         setFlagPV(isEvenNumberBits(i));
-        setFlagN(false);
+        resetFlagN();
         setFlagXY(i);
     }
 
@@ -5476,7 +5494,7 @@ class Z80
         reg.pair.B--;
         reg.pair.F = 0;
         setFlagC(isFlagC());
-        setFlagN(true);
+        setFlagN();
         setFlagZ(reg.pair.B == 0);
         setFlagXY(reg.pair.B);
         setFlagS(reg.pair.B & 0x80);
@@ -5629,9 +5647,9 @@ class Z80
         setFlagS(reg.pair.A & 0x80);
         setFlagXY(reg.pair.A);
         setFlagZ(reg.pair.A == 0);
-        setFlagH(false);
+        resetFlagH();
         setFlagPV(isEvenNumberBits(reg.pair.A));
-        setFlagN(false);
+        resetFlagN();
         consumeClock(2);
     }
 
@@ -5658,9 +5676,9 @@ class Z80
         setFlagS(reg.pair.A & 0x80);
         setFlagXY(reg.pair.A);
         setFlagZ(reg.pair.A == 0);
-        setFlagH(false);
+        resetFlagH();
         setFlagPV(isEvenNumberBits(reg.pair.A));
-        setFlagN(false);
+        resetFlagN();
         consumeClock(2);
     }